Finalizado a conexão com fireBase, criado validação se o usuário existe, criado arquivo MonthSelector que é responsavel pela lógica dos meses, criado arquivo login e cadastro de usuário, criado a coluna de createdAt, foi arrumado o arquivo productList que mostrava os cards dos produtos com height de 100%

+++ src/service/ProductService.js
@@ -1,29 +1,53 @@
 import { getDB } from '../database/db';
+import { auth } from '../service/AuthService';
+
+import { Alert } from 'react-native';
 
 export async function getProducts() {
   const db = getDB();
+  const userId = auth.currentUser?.uid; // Pegar o ID do usuário logado
+
+  if(!userId) {
+    Alert.alert("Faça login para ver seus produtos.");
+    return [];
+  }
+
   const result = await db.getAllAsync(
-    'SELECT * FROM products ORDER BY id DESC'
+    'SELECT * FROM products  WHERE userId = ? ORDER BY id DESC', [userId]
   );
-  console.log('📦 PRODUTOS DO BANCO:', result);
+
+  // console.log('📦 PRODUTOS DO BANCO:', userId, result);
   return result;
 }
 
 export async function createProduct(name, price, quantity, category) {
   const db = getDB();
+  const user = auth.currentUser;
+  const createdAt = new Date().toISOString(); // Pega a data e hora atual
+
+  if(!user){
+    Alert.alert("Usuário não identificado!");
+    return
+  }
+
+  const userId = user.uid;
 
   // Converta para garantir que o banco receba números
   const numericPrice = parseFloat(price);
   const numericQuantity = parseInt(quantity, 10);
 
-  console.log('🟡 CREATE PRODUCT:', name, price, quantity, category);
+  //DEBUG
+  // console.log('🟡 CREATE PRODUCT:', name, price, quantity, category, userId);
 
   try {
     const result = await db.runAsync(
-      'INSERT INTO products (name, price, quantity, category) VALUES (?, ?, ?, ?)',
-      [name, numericPrice, numericQuantity, category]
+      'INSERT INTO products (name, price, quantity, category, userId, createdAt) VALUES (?, ?, ?, ?, ?, ?)',
+      [name, numericPrice, numericQuantity, category, userId, createdAt]
     );
-    console.log('✅ SUCESSO NO BANCO! ID Gerado:', result.lastInsertRowId);
+
+    // DEBUG
+    // console.log('✅ SUCESSO NO BANCO! ID Gerado:', result.lastInsertRowId);
+    
     return result; 
   } catch (error) {
     console.error('❌ ERRO REAL AO INSERIR:', error);
@@ -33,13 +57,15 @@ export async function createProduct(name, price, quantity, category) {
 
 export async function deleteProduct(id) {
   const db = getDB();
-  await db.runAsync('DELETE FROM products WHERE id = ?', [id]);
+  const userId = auth.currentUser?.uid;
+  await db.runAsync('DELETE FROM products WHERE id = ? AND userId = ?', [id, userId]);
 }
 
 export async function updateProduct(id, name, price, quantity, category) {
   const db = getDB();
+  const userId = auth.currentUser?.uid;
   await db.runAsync(
-    'UPDATE products SET name = ?, price = ?, quantity = ?, category = ? WHERE id = ?',
-    [name, price, quantity, category, id]
+    'UPDATE products SET name = ?, price = ?, quantity = ?, category = ? WHERE id = ? AND userId = ?',
+    [name, price, quantity, category, id, userId]
   );
 }
